refactor admin page routing
